Added support for variant quests to drop their independent rewards as defined in their JSON. Renamed variant id to be more explicit within the schema.

+++ Arrowgene.Ddon.GameServer/Handler/QuestQuestProgressHandler.cs
@@ -75,7 +75,7 @@ namespace Arrowgene.Ddon.GameServer.Handler
 
                         if (quest.IsVariantQuest)
                         {
-                            if (!Server.Database.InsertQuestProgress(memberClient.Character.CommonId, quest.QuestId, quest.QuestType, 0, (uint)quest.VariantId))
+                            if (!Server.Database.InsertQuestProgress(memberClient.Character.CommonId, quest.QuestId, quest.QuestType, 0, quest.VariantId))
                             {
                                 Logger.Error($"Failed to insert progress for the quest {quest.QuestId}");
                             }
